dahboard관련 api 3개 개발, 테스트코드x

+++ src/main/java/com/developlife/reviewtwits/message/response/project/ProductStatisticsResponse.java
@@ -4,11 +4,11 @@ import lombok.Builder;
 
 public record ProductStatisticsResponse(
         String productName,
-        Integer visitCount,
-        Integer reviewCount,
-        Integer mainAge,
+        Long visitCount,
+        Long reviewCount,
+        Long mainAge,
         String mainGender,
-        Float averageScore
+        Double averageScore
 ) {
     @Builder
     public ProductStatisticsResponse {
